moved HTML template back to own files
