+++ sketches/timinggatecontroller/htmltemplate.h
@@ -1,4 +1,13 @@
 #ifndef _htmltemplate_h
 #define _htmltemplate_h
 
+#include <Arduino.h>    // for PROGMEM
+
+namespace HTML 
+{
+
+extern const char PROGMEM index[]; 
+
+};
+
 #endif
